feat: set upper limit for jobs fetched
If there are too many pages program with fetch up to MAX_PAGE pages.
This upper limit is added in order to sending many requests for
unnecessary data.

Diagnostic request data is also improved by outputing related data
to the request and coupling it together (separating output  of each
request)